+++ src/lib/Layout/PanelSideBarLayout/PanelSideBar/Context/PanelSideBarContext.tsx
@@ -69,6 +69,15 @@ export interface PanelSideBarContextProps<TPanelItem> {
    * The context theme
    */
   theme?: "light";
+ /**
+   * Boolean indicating if you want to render first items level as icons or directly as menu entries
+   */
+  renderFirstItemsLevelAsTiles?: boolean;
+
+   /**
+   * Boolean indicating if you want to render first level items as links or as button
+   */
+  renderTilesAsLinks?: boolean;
 }
 
 export const PanelSideBarContext = createContext<PanelSideBarContextProps<any> | null>(null);
@@ -86,6 +95,8 @@ export interface PanelSideBarMenuProviderProps<TPanelItem>
     | "topBarLeftCustomItems"
     | "localItems"
     | "theme"
+    | "renderFirstItemsLevelAsTiles"
+    | "renderTilesAsLinks"
   > {
   /**
    * The children elements.
@@ -105,6 +116,8 @@ export const PanelSideBarProvider = <TPanelItem,>(props: PanelSideBarMenuProvide
     userDropDownMenuToggle,
     topBarRightCustomItems,
     topBarLeftCustomItems,
+    renderFirstItemsLevelAsTiles = true,
+    renderTilesAsLinks = false,
     theme = "light",
   } = props;
 
@@ -155,6 +168,8 @@ export const PanelSideBarProvider = <TPanelItem,>(props: PanelSideBarMenuProvide
         topBarLeftCustomItems,
         brand,
         theme,
+        renderFirstItemsLevelAsTiles,
+        renderTilesAsLinks,
       }}
     >
       {children}