+++ src/lib/Layout/PanelSideBarLayout/PanelSideBarLayout.tsx
@@ -37,6 +37,7 @@ export const PanelSideBarLayout = (props: PanelSideBarLayoutProps) => {
     userDropDownMenuToggle,
     topBarRightCustomItems = [],
     topBarLeftCustomItems = [],
+    renderFirstItemsLevelAsTiles,
   } = usePanelSideBarContext();
 
   return (
@@ -75,8 +76,8 @@ export const PanelSideBarLayout = (props: PanelSideBarLayoutProps) => {
         </Nav>
       </nav>
 
-      <section className={classNames({ toggled: !isOpen })}>
-        <PanelSideBar toggledSidebar={isOpen} />
+      <section className={`${renderFirstItemsLevelAsTiles ? "section-tiles" : "section-no-tiles" } ${classNames({ toggled: !isOpen })}`}>
+        <PanelSideBar />
         {collapsible && <PanelSideBarToggle onClick={toggleSidebar} toggled={!isOpen} />}
         <PanelSideBarLayoutContent footer={footer ?? contextFooter}>{children}</PanelSideBarLayoutContent>
       </section>